requirements.txt
gspread
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
feedparser
openai
python-dotenv
schedule
fastapi
uvicorn
tavily-python
cloudinary
httpx
python-jose[cryptography]
bcrypt
cryptography
groq
pydantic
python-multipart
python-jose
jwt
# For vision
pillow
